requirements.txt
Flask>=2.0
pandas>=1.0
numpy>=1.20
Flask-CORS

# Data Science & Analysis
scipy>=1.7.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Data Manipulation
openpyxl>=3.0.7
xlrd>=2.0.1
requests>=2.25.0

# Image Processing
Pillow>=8.0.0

# Natural Language Processing
nltk>=3.7

# Web Scraping
beautifulsoup4>=4.10.0
lxml>=4.7.0

# Utilities
python-dateutil>=2.8.2
pytz>=2021.3
tqdm>=4.62.0
